sc2: renamed test files; added a test for campaign size limit matching the number of missions
