+++ worlds/sc2/test/test_options.py
@@ -1,0 +1,10 @@
+import unittest
+from .test_base import Sc2TestBase
+from .. import Options, MissionTables
+
+class TestOptions(unittest.TestCase):
+    def test_campaign_size_option_max_matches_number_of_missions(self):
+        self.assertEqual(Options.MaximumCampaignSize.range_end, len(MissionTables.SC2Mission))
+
+if __name__ == '__main__':
+    unittest.main(verbosity=2)
